Allow publishing despite vulnerable dependency.
Until `socket.io` publishes a new version that depends on a newer version
of `engine.io` that depends on a newer version of `accepts` that depends
on a newer version of `negotiator` which fixed the security
vulnerability, I cannot publish this package to `npm`.

This is probably a very bad idea for production apps, but this is a
development tool.